Ingestion/requirements.txt
attrs==25.3.0
boto3==1.40.11
botocore==1.40.11
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
referencing==0.36.2
rpds-py==0.27.0
s3transfer==0.13.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
